feat: add toDOM adaptor to communicate with wysiwyg schema
