+++ apps/editor/src/editor.ts
@@ -31,6 +31,7 @@ import { invokePlugins, getPluginInfo } from './pluginHelper';
 // @ts-ignore
 import { ToastMark } from '@toast-ui/toastmark';
 import isString from 'tui-code-snippet/type/isString';
+import { WwToDOM } from './wysiwyg/adaptor/wwToDOM';
 
 /**
  * ToastUI Editor
@@ -175,6 +176,7 @@ class ToastUIEditor {
       customParser: parser,
       frontMatter
     };
+    const wwToDOM = new WwToDOM(linkAttribute, rendererOptions.customHTMLRenderer);
 
     if (this.options.hooks) {
       forEachOwnProperties(this.options.hooks, (fn, key) => this.addHook(key, fn));
@@ -212,7 +214,11 @@ class ToastUIEditor {
       highlight: this.options.previewHighlight
     });
 
-    this.wwEditor = new WysiwygEditor(this.layout.getWwEditorContainerEl(), this.eventEmitter);
+    this.wwEditor = new WysiwygEditor(
+      this.layout.getWwEditorContainerEl(),
+      this.eventEmitter,
+      wwToDOM
+    );
 
     this.convertor = new Convertor(this.wwEditor.getSchema());
