+++ apps/editor/types/prosemirror-model.d.ts
@@ -4,6 +4,7 @@ declare module 'prosemirror-model' {
   export interface Fragment {
     textBetween(from: number, to: number, blockSeparator?: string, leafText?: string): string;
     findIndex(pos: number, round?: number): { index: number; offset: number };
+    findDiffEnd(other: ProsemirrorNode | Fragment): { a: number; b: number } | null | undefined;
   }
 
   export interface ProsemirrorNode extends Model.Node {
